DE-149 add in UI for jasper report dialog along with reporting files

+++ eppt-jasper-engine/src/test/java/gov/ca/water/eppt/jasperengine/TestJasperReportRunner.java
@@ -46,11 +46,14 @@ public class TestJasperReportRunner
 		Path output = Paths.get("Output.pdf");
 		try
 		{
-			Files.delete(output);
+			if(output.toFile().exists())
+			{
+				Files.delete(output);
+			}
 		}
 		catch(NoSuchFileException ex)
 		{
-			LOGGER.log(Level.FINE, "File already deleted", ex);
+			LOGGER.log(Level.SEVERE, "File already deleted", ex);
 		}
 		jasperReportRunner.runReportWithOutputFile(output, jrxmlPath);
 		assertTrue(output.toFile().exists(), "Output PDF should be generated");
